icu_collator

Module comparison

This module holds the Collator struct whose compare_impl() contains the comparison of collation element sequences.

Macros

collation_elements 🔒

compare 🔒

hangul_syllable_compare 🔒

negdivmod 🔒

quick_primary_compare 🔒

Structs

AnyQuaternaryAccumulator 🔒

Collator

Compares strings according to culturally-relevant ordering.

CollatorBorrowed

Compares strings according to culturally-relevant ordering, borrowed version.

CollatorPreferences

The preferences for collation.

LocaleSpecificDataHolder 🔒

Holder struct for payloads that are locale-dependent. (For code reuse between owned and borrowed cases.)

SortKeyLevel 🔒

TooSmall

Error indicating that a CollationKeySink with limited space ran out of space.

Constants

CASE_LEVEL_FLAG 🔒

CASE_LOWER_FIRST_COMMON 🔒

CASE_UPPER_FIRST_COMMON 🔒

COMMON_BYTE 🔒

Default secondary/tertiary weight lead byte.

COMMON_WEIGHT16 🔒

LEVEL_MASKS 🔒

LEVEL_SEPARATOR_BYTE 🔒

MERGE_SEPARATOR 🔒

Merge-sort-key separator.

MERGE_SEPARATOR_BYTE 🔒

MERGE_SEPARATOR_PRIMARY 🔒

PRIMARY_COMPRESSION_HIGH_BYTE 🔒

Primary compression high terminator.

PRIMARY_COMPRESSION_LOW_BYTE 🔒

Primary compression low terminator, must be greater than MERGE_SEPARATOR_BYTE.

PRIMARY_LEVEL_FLAG 🔒

QUATERNARY_LEVEL_FLAG 🔒

QUAT_COMMON 🔒

QUAT_SHIFTED_LIMIT_BYTE 🔒

SECONDARY_LEVEL_FLAG 🔒

SEC_COMMON 🔒

SLOPE_LEAD_2 🔒

SLOPE_LEAD_3 🔒

SLOPE_MAX 🔒

SLOPE_MIDDLE 🔒

SLOPE_MIN 🔒

SLOPE_REACH_NEG_1 🔒

SLOPE_REACH_NEG_2 🔒

SLOPE_REACH_NEG_3 🔒

SLOPE_REACH_POS_1 🔒

SLOPE_REACH_POS_2 🔒

SLOPE_REACH_POS_3 🔒

SLOPE_SINGLE 🔒

SLOPE_START_NEG_2 🔒

SLOPE_START_NEG_3 🔒

SLOPE_START_POS_2 🔒

SLOPE_START_POS_3 🔒

SLOPE_TAIL_COUNT 🔒

TERTIARY_LEVEL_FLAG 🔒

TER_LOWER_FIRST_COMMON 🔒

TER_ONLY_COMMON 🔒

TER_UPPER_FIRST_COMMON 🔒

WEIGHT_HIGH 🔒

WEIGHT_LOW 🔒

WEIGHT_MAX_COUNT 🔒

WEIGHT_MIDDLE 🔒

Traits

CollationKeySink

A std::io::Write-like trait for writing to a buffer-like object.

Functions

in_inclusive_range16 🔒

true iff i is greater or equal to start and less or equal to end.

split_prefix 🔒

Finds the identical prefix of left and right containing guaranteed well-format UTF-8.

split_prefix_latin1 🔒

Finds the identical prefix of left and right containing Latin1.

split_prefix_latin1_utf16 🔒

Finds the identical prefix of left containing Latin1 and right containing potentially ill-formed UTF-16.

split_prefix_u8 🔒

Finds the identical prefix of left and right containing potentially ill-formed UTF-8, while avoiding splitting a UTF-8 byte sequence. In case of ill-formed UTF-8, the prefix is not guaranteed to be maximal.

split_prefix_u16 🔒

Finds the identical prefix of left and right containing potentially ill-formed UTF-16, while avoiding splitting a well-formed surrogate pair. In case of ill-formed UTF-16, the prefix is not guaranteed to be maximal.

write_diff 🔒

write_identical_level 🔒

Type Aliases

NormTrie 🔒