icu_collator::comparison

Trait CollationKeySink

pub trait CollationKeySink {
    type Error;
    type State;
    type Output;

    // Required methods
    fn write(
        &mut self,
        state: &mut Self::State,
        buf: &[u8],
    ) -> Result<(), Self::Error>;
    fn finish(
        &mut self,
        state: Self::State,
    ) -> Result<Self::Output, Self::Error>;

    // Provided method
    fn write_byte(
        &mut self,
        state: &mut Self::State,
        b: u8,
    ) -> Result<(), Self::Error> { ... }
}

A std::io::Write-like trait for writing to a buffer-like object.

(This crate does not have access to std.)

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement or call methods on this trait unless you are prepared for things to occasionally break.
Graduation tracking issue: issue #7178.

✨ Enabled with the unstable Cargo feature.

Required Associated Types

type Error

The type of error the sink may return.

type State

An intermediate state object used by the sink, which must implement Default.

type Output

A result value indicating the final state of the sink (e.g. a number of bytes written).

Required Methods

fn write( &mut self, state: &mut Self::State, buf: &[u8], ) -> Result<(), Self::Error>

Writes a buffer into the writer.

fn finish(&mut self, state: Self::State) -> Result<Self::Output, Self::Error>

Finalize any internal sink state (perhaps by flushing a buffer) and return the final output value.

Provided Methods

fn write_byte( &mut self, state: &mut Self::State, b: u8, ) -> Result<(), Self::Error>

Write a single byte into the writer.