Module icu_collator::options

This module contains the types that are part of the API for setting the options for the collator.

Structs

CollatorOptions

Options settable by the user of the API.

CollatorOptionsBitField 🔒

ResolvedCollatorOptions

The resolved (actually used) options used by the collator.

Enums

AlternateHandling

What to do about characters whose comparison level can be varied dynamically.

BackwardSecondLevel

Whether second level compares the last accent difference instead of the first accent difference.

CaseFirst

Treatment of case. (Large and small kana differences are treated as case differences.)

CaseLevel

Whether to distinguish case in sorting, even for sorting levels higher than tertiary, without having to use tertiary level just to enable case level differences.

MaxVariable

What characters get shifted to the quaternary level with AlternateHandling::Shifted.

Numeric

When set to On, any sequence of decimal digits is sorted at a primary level according to the numeric value.

Strength

The collation strength that indicates how many levels to compare. If an earlier level isn’t equal, the earlier level is decisive. If the result is equal on a level, but the strength is higher, the comparison proceeds to the next level.