icu_collator::options

Enum Strength

#[non_exhaustive]

#[repr(u8)]
pub enum Strength {
    Primary = 0,
    Secondary = 1,
    Tertiary = 2,
    Quaternary = 3,
    Identical = 7,
}

The collation strength that indicates how many levels to compare. The primary level considers base letters, i.e. ‘a’ and ‘b’ are unequal but ‘E’ and ‘é’ are equal, with further levels dealing with distinctions such as accents and case.

Note that what constitutes a base letter depends on the language and not on Unicode character properties. For example, ‘ö’ is analyzed as a base letter for various languages (e.g. Estonian, Finnish, Icelandic, Swedish, and Turkish), so it is distinct from ‘o’ even on the primary level in such languages. Inputs that are equal in fold case (as tailored for e.g. Turkish) are expected to be equal on the primary level. For example, “ß” is primary-equal with “ss”. Characters that are graphically ligature-like can be primary-equal with what they appear to be ligatures of. For example, in the root collation (but not in e.g. Danish and Norwegian) “æ” is primary-equal with “ae”.

If an earlier level isn’t equal, the earlier level is decisive. If the result is equal on a level, but the strength is higher, the comparison proceeds to the next level.

Note that lowering the strength means that more user-perceptible differences compare as equal. This may make sense when sorting more complex structures where the string to be compared is just one field, and ties between strings that differ only in case, accent, or similar are resolved by comparing some secondary field in the larger structure to be sorted.

Therefore, if the sort is just a string sort without some other field for resolving ties, lowering the strength means that factors that don’t make sense to the user (such as the order of items prior to sorting with a stable sort algorithm or the internal details of a sorting algorithm that doesn’t provide the stability property) affect the relative order of strings that do have user-perceptible differences particularly in accents or case.

Lowering the strength is less of a perfomance optimization that it may seem directly from the above description. As described above, in the case of identical strings to be compared, the algorithm has to work though all the levels included in the strength without an early exit. However, this collator implements an identical prefix optimization, which examines the code units of the strings to be compared to skip the identical prefix before starting the actual collation algorithm. When the strings to be compared are identical on the byte level, they are found to be equal without the actual collation algorithm running at all! Therefore, the strength setting only has an effect (whether order effect or performance effect) for comparisons where the strings to be compared are not equal on the byte level but are equal on the primary level/strength. The common cases are that a comparison is decided on the primary level or the strings are byte equal, which narrows the performance effect of lowering the strength setting.

Note: The bit layout of CollatorOptionsBitField requires Strength to fit in 3 bits.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Primary = 0

Compare only on the level of base letters. This level corresponds to the ECMA-402 sensitivity “base” with CaseLevel::Off (the default for CaseLevel) and to ECMA-402 sensitivity “case” with CaseLevel::On.

use icu::collator::{options::*, *};

let mut options = CollatorOptions::default();
options.strength = Some(Strength::Primary);
let collator = Collator::try_new(Default::default(), options).unwrap();
assert_eq!(collator.compare("E", "é"), core::cmp::Ordering::Equal);

Secondary = 1

Compare also on the secondary level, which corresponds to diacritics in scripts that use them. This level corresponds to the ECMA-402 sensitivity “accent”.

use icu::collator::{options::*, *};

let mut options = CollatorOptions::default();
options.strength = Some(Strength::Secondary);
let collator = Collator::try_new(Default::default(), options).unwrap();
assert_eq!(collator.compare("E", "e"), core::cmp::Ordering::Equal);
assert_eq!(collator.compare("e", "é"), core::cmp::Ordering::Less);
assert_eq!(collator.compare("あ", "ア"), core::cmp::Ordering::Equal);
assert_eq!(collator.compare("ァ", "ア"), core::cmp::Ordering::Equal);
assert_eq!(collator.compare("ア", "ｱ"), core::cmp::Ordering::Equal);

Tertiary = 2

Compare also on the tertiary level. By default, if the separate case level is disabled, this corresponds to case for bicameral scripts. This level distinguishes Hiragana and Katakana. This also captures other minor differences, such as half-width vs. full-width when the Japanese tailoring isn’t in use.

This is the default comparison level and appropriate for most scripts. This level corresponds to the ECMA-402 sensitivity “variant”.

use icu::collator::{*, options::*};
use icu::locale::locale;

let mut options = CollatorOptions::default();
options.strength = Some(Strength::Tertiary);
let collator =
  Collator::try_new(Default::default(),
                    options).unwrap();
assert_eq!(collator.compare("E", "e"),
           core::cmp::Ordering::Greater);
assert_eq!(collator.compare("e", "é"),
           core::cmp::Ordering::Less);
assert_eq!(collator.compare("あ", "ア"),
           core::cmp::Ordering::Less);
assert_eq!(collator.compare("ァ", "ア"),
           core::cmp::Ordering::Less);
assert_eq!(collator.compare("ア", "ｱ"),
           core::cmp::Ordering::Less);
assert_eq!(collator.compare("e", "ｅ"), // Full-width e
           core::cmp::Ordering::Less);

let ja_collator =
  Collator::try_new(locale!("ja").into(), options).unwrap();
assert_eq!(ja_collator.compare("E", "e"),
           core::cmp::Ordering::Greater);
assert_eq!(ja_collator.compare("e", "é"),
           core::cmp::Ordering::Less);
assert_eq!(ja_collator.compare("あ", "ア"),
           core::cmp::Ordering::Equal); // Unlike root!
assert_eq!(ja_collator.compare("ァ", "ア"),
           core::cmp::Ordering::Less);
assert_eq!(ja_collator.compare("ア", "ｱ"),
           core::cmp::Ordering::Equal); // Unlike root!
assert_eq!(ja_collator.compare("e", "ｅ"), // Full-width e
           core::cmp::Ordering::Equal); // Unlike root!

Quaternary = 3

Compare also on the quaternary level. For Japanese, Hiragana and Katakana are distinguished at the quaternary level. Also, if AlternateHandling::Shifted is used, the collation elements whose level gets shifted are shifted to this level.

use icu::collator::{*, options::*};
use icu::locale::locale;

let mut options = CollatorOptions::default();
options.strength = Some(Strength::Quaternary);

let ja_collator =
  Collator::try_new(locale!("ja").into(), options).unwrap();
assert_eq!(ja_collator.compare("あ", "ア"),
           core::cmp::Ordering::Less);
assert_eq!(ja_collator.compare("ア", "ｱ"),
           core::cmp::Ordering::Equal);
assert_eq!(ja_collator.compare("e", "ｅ"), // Full-width e
           core::cmp::Ordering::Equal);

// Even this level doesn't distinguish everything,
// e.g. Hebrew cantillation marks are still ignored.
let collator =
  Collator::try_new(Default::default(),
                    options).unwrap();
assert_eq!(collator.compare("דחי", "דחי֭"),
           core::cmp::Ordering::Equal);

Identical = 7

Compare the NFD form by code point order as the quinary level. This level makes the comparison slower and should not be used in the general case. However, it can be used to distinguish full-width and half-width forms when the Japanese tailoring is in use and to distinguish e.g. Hebrew cantillation markse. Use this level if you need JIS X 4061-1996 compliance for Japanese on the level of distinguishing full-width and half-width forms.

use icu::collator::{*, options::*};
use icu::locale::locale;

let mut options = CollatorOptions::default();
options.strength = Some(Strength::Identical);

let ja_collator =
  Collator::try_new(locale!("ja").into(), options).unwrap();
assert_eq!(ja_collator.compare("ア", "ｱ"),
           core::cmp::Ordering::Less);
assert_eq!(ja_collator.compare("e", "ｅ"), // Full-width e
           core::cmp::Ordering::Less);

let collator =
  Collator::try_new(Default::default(),
                    options).unwrap();
assert_eq!(collator.compare("דחי", "דחי֭"),
           core::cmp::Ordering::Less);