icu_collator::provider

Struct CollationMetadata

pub struct CollationMetadata {
    pub bits: u32,
}

Each non-alias collation that the data provider knows about explicitly has an data entry at least for this struct.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
bits: u32

See the mask constants in the impl block for the bit layout. The other bits are ignored: They could be from the future if their semantics such that old code may ignore them.

Note: At present, it’s bogus for the bit for “upper first” to be set if “case first” isn’t also set. However, the methods handle this case gracefully, so there is no need for invariant validation.

Implementations

impl CollationMetadata

const MAX_VARIABLE_MASK: u32 = 0b11

const TAILORED_MASK: u32

const TAILORED_DIACRITICS_MASK: u32

const REORDERING_MASK: u32

const LITHUANIAN_DOT_ABOVE_MASK: u32

const BACWARD_SECOND_LEVEL_MASK: u32

const ALTERNATE_SHIFTED_MASK: u32

const CASE_FIRST_MASK: u32

const UPPER_FIRST_MASK: u32

pub(crate) fn max_variable(self) -> MaxVariable

pub(crate) fn tailored(self) -> bool

pub(crate) fn tailored_diacritics(self) -> bool

Vietnamese and Ewe

pub(crate) fn lithuanian_dot_above(self) -> bool

Lithuanian

pub(crate) fn backward_second_level(self) -> bool

Canadian French

pub(crate) fn reordering(self) -> bool

pub(crate) fn alternate_shifted(self) -> bool

Thai

pub(crate) fn case_first(self) -> CollationCaseFirst