Struct icu_collator::Collator

pub struct Collator { /* private fields */ }

Compares strings according to culturally-relevant ordering.

Implementations

impl Collator

pub fn try_new_unstable<D>(
data_provider: &D,
locale: &DataLocale,
options: CollatorOptions
) -> Result<Self, CollatorError> where
D: DataProvider<CollationSpecialPrimariesV1Marker> + DataProvider<CollationDataV1Marker> + DataProvider<CollationDiacriticsV1Marker> + DataProvider<CollationJamoV1Marker> + DataProvider<CollationMetadataV1Marker> + DataProvider<CollationReorderingV1Marker> + DataProvider<CanonicalDecompositionDataV1Marker> + DataProvider<CanonicalDecompositionTablesV1Marker> + ?Sized,

Instantiates a collator for a given locale with the given options

pub fn try_new_with_any_provider(
provider: &impl AnyProvider + ?Sized,
locale: &DataLocale,
options: CollatorOptions
) -> Result<Self, CollatorError>

Create a new instance using an AnyProvider.

See also: Self::try_new_unstable

pub fn try_new_with_buffer_provider(
provider: &impl BufferProvider + ?Sized,
locale: &DataLocale,
options: CollatorOptions
) -> Result<Self, CollatorError>

Create a new instance using a BufferProvider. Enabled with the "serde" feature.

See also: Self::try_new_unstable

pub fn compare_utf16(&self, left: &[u16], right: &[u16]) -> Ordering

Compare potentially-invalid UTF-16 slices. Unpaired surrogates are compared as if each one was a REPLACEMENT CHARACTER.

pub fn compare(&self, left: &str, right: &str) -> Ordering

Compare guaranteed-valid UTF-8 slices.

pub fn compare_utf8(&self, left: &[u8], right: &[u8]) -> Ordering

Compare potentially-valid UTF-8 slices. Invalid input is compared as if errors had been replaced with REPLACEMENT CHARACTERs according to the WHATWG Encoding Standard.