Struct icu_collator::CollatorOptions

pub struct CollatorOptions(_);

Options settable by the user of the API.

See the spec.

The setters take an Option so that None can be used to go back to default.

Options

Strength

This is the BCP47 key ks. The default is Strength::Tertiary.

Alternate Handling

This is the BCP47 key ka. Note that ShiftTrimmed and Blanked are unimplemented. The default is AlternateHandling::NonIgnorable, except for Thai, whose default is AlternateHandling::Shifted.

Case Level

See the spec. This is the BCP47 key kc. The default is false (off).

Case First

See the spec. This is the BCP47 key kf. Three possibilities: CaseFirst::Off (default, except for Danish and Maltese), CaseFirst::Lower, and CaseFirst::Upper (default for Danish and Maltese).

Backward second level

Compare the second level in backward order. This is the BCP47 key kb. kb is prohibited by ECMA 402. The default is false (off), except for Canadian French.

Numeric

This is the BCP47 key kn. When set to true (on), any sequence of decimal digits (General_Category = Nd) is sorted at a primary level accoding to the numeric value. The default is false (off).

Unsupported BCP47 options

Reordering (BCP47 kr) currently cannot be set via the API and is implied by the locale of the collation. kr is probihibited by ECMA 402.

Normalization is always enabled and cannot be turned off. Therefore, there is no option corresponding to BCP47 kk. kk is prohibited by ECMA 402.

Hiragana quaternary handling is part of the strength for the Japanese tailoring. The BCP47 key kh is unsupported. kh is deprecated and prohibited by ECMA 402.

Variable top (BCP47 vt) is unsupported (use Max Variable instead). vt is deprecated and prohibited by ECMA 402.

Implementations

impl CollatorOptions

pub const fn new() -> Self

Create a new CollatorOptions with the defaults.

pub fn strength(&self) -> Strength

This is the BCP47 key ks.

pub fn set_strength(&mut self, strength: Option<Strength>)

This is the BCP47 key ks. See the enum for examples.

pub fn max_variable(&self) -> MaxVariable

The maximum character class that AlternateHandling::Shifted applies to.

pub fn set_max_variable(&mut self, max_variable: Option<MaxVariable>)

The maximum character class that AlternateHandling::Shifted applies to. See the enum for examples.

pub fn alternate_handling(&self) -> AlternateHandling

Whether certain characters are moved from the primary level to the quaternary level.

pub fn set_alternate_handling(
&mut self,
alternate_handling: Option<AlternateHandling>
)

Whether certain characters are moved from the primary level to the quaternary level. See the enum for examples.

pub fn case_level(&self) -> bool

Whether there’s a dedicated case level.

pub fn set_case_level(&mut self, case_level: Option<bool>)

Whether there’s a dedicated case level. If true, detaches the case aspect of the tertiary level and inserts it between the secondary and tertiary levels. Can be combined with the primary-only strength. Setting this to true with Strength::Primary corresponds to the ECMA-402 sensitivity “case”.

See the ICU guide.

use icu_collator::*;

let data_provider = icu_testdata::get_provider();
let mut options = CollatorOptions::new();
options.set_strength(Some(Strength::Tertiary));
let tertiary =
  Collator::try_new(&data_provider,
                    &Default::default(),
                    options).unwrap();
// The first string starts with full-width a
assert_eq!(tertiary.compare("ａa", "aA"),
           core::cmp::Ordering::Greater);

options.set_case_level(Some(true));
let tertiary_and_case =
  Collator::try_new(&data_provider,
                    &Default::default(),
                    options).unwrap();
// The first string starts with full-width a
// TODO!!!!

pub fn case_first(&self) -> CaseFirst

Whether case is the most significant part of the tertiary level.

See the ICU guide.

pub fn set_case_first(&mut self, case_first: Option<CaseFirst>)

Whether case is the most significant part of the tertiary level.

See the ICU guide.

pub fn backward_second_level(&self) -> bool

Whether second level compares the last accent difference instead of the first accent difference.

pub fn set_backward_second_level(&mut self, backward_second_level: Option<bool>)

Whether second level compares the last accent difference instead of the first accent difference.

pub fn numeric(&self) -> bool

Whether sequences of decimal digits are compared according to their numeric value.

pub fn set_numeric(&mut self, numeric: Option<bool>)

Whether sequences of decimal digits are compared according to their numeric value.

pub fn set_defaults(&mut self, other: CollatorOptions)

For options left as defaults in this CollatorOptions, set the value from other. Values taken from other are marked as explicitly set if they were explicitly set in other.